More tests around getting position
