+++ errors.go
@@ -11,6 +11,9 @@ var (
 	ErrInvalidMsgInterval                            = errors.New("MsgInterval cannot be less than 2")
 	ErrSubscriberNeedsCategoryOrStream               = errors.New("Subscriber needs at least one of category or stream to be set upon creation")
 	ErrSubscriberIDCannotBeEmpty                     = errors.New("Subscriber ID cannot be nil")
+	ErrSubscriberIDCannotContainPlus                 = errors.New("Subscriber ID cannot contain a plus")
+	ErrSubscriberIDCannotContainHyphen               = errors.New("Subscriber ID cannot contain a hyphen")
+	ErrPositionVersionMissing                        = errors.New("Subscriber cannot use non-versioned positionMsg object")
 	ErrSubscriberNeedsAtLeastOneMessageHandler       = errors.New("Subscriber needs at least one handler upon creation")
 	ErrSubscriberCannotSubscribeToMultipleStreams    = errors.New("Subscribers can only subscribe to one stream")
 	ErrSubscriberCannotSubscribeToMultipleCategories = errors.New("Subscribers can only subscribe to one category")
@@ -33,7 +36,8 @@ var (
 	ErrInvalidMessageCategory                        = errors.New("Hyphens are not allowed in category names")
 	ErrInvalidCommandStream                          = errors.New("Hyphens are not allowed in command stream name")
 	ErrInvalidEventStream                            = errors.New("Hyphens are not allowed in event stream name")
-	ErrInvalidPositionStream                         = errors.New("Hyphens are not allowed in position stream name")
+	ErrInvalidSubscriberID                           = errors.New("Hyphens and plusses are not allowed in subscriber ID")
+	ErrInvalidPositionStream                         = errors.New("Position stream expects to have a single plus diving the subscriber ID and the word 'position'")
 	ErrMissingMessageCategoryID                      = errors.New("All messages require a category ID")
 	ErrMissingMessageData                            = errors.New("Messages payload must not be nil")
 	ErrUnserializableData                            = errors.New("Message data could not be encoded as json")
